--- app/cdap/components/AddTransformation/index.tsx
@@ -1,121 +1,0 @@
-/*
- * Copyright © 2022 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-import DrawerWidget from 'components/DrawerWidget';
-import T from 'i18n-react';
-import React, { Fragment, useState, useEffect } from 'react';
-import SelectColumnsList from 'components/AddTransformation/SelectColumnsList';
-import {
-  IAddTransformationProps,
-  IHeaderNamesList,
-  IMultipleSelectedFunctionDetail,
-  IDataQualityItem,
-} from 'components/AddTransformation/types';
-import { getDataQuality } from 'components/common/DataQualityCircularProgressBar/utils';
-import {
-  multipleColumnSelected,
-  ADD_TRANSFORMATION_PREFIX,
-} from 'components/AddTransformation/constants';
-import {
-  AddTransformationBodyWrapper,
-  AddTransformationWrapper,
-} from 'components/common/BoxContainer';
-import { AddTransformationButton } from 'components/common/ButtonWidget';
-
-export default function ({
-  transformationDataType,
-  transformationName,
-  columnsList,
-  missingItemsList,
-  onCancel,
-}: IAddTransformationProps) {
-  const [columnsPopup, setColumnsPopup] = useState<boolean>(true);
-  const [selectedColumns, setSelectedColumns] = useState<IHeaderNamesList[]>([]);
-  const [dataQualityValue, setDataQualityValue] = useState<IDataQualityItem[]>([]);
-
-  const closeSelectColumnsPopup = () => {
-    setColumnsPopup(false);
-    onCancel();
-  };
-
-  const closeSelectColumnsPopupWithoutColumn = () => {
-    setColumnsPopup(false);
-    setSelectedColumns([]);
-    onCancel();
-  };
-
-  useEffect(() => {
-    const getPreparedDataQuality: IDataQualityItem[] = getDataQuality(
-      missingItemsList,
-      columnsList
-    );
-    setDataQualityValue(getPreparedDataQuality);
-  }, []);
-
-  const enableDoneButton = () => {
-    if (
-      multipleColumnSelected.filter(
-        (functionNameDetail: IMultipleSelectedFunctionDetail) =>
-          functionNameDetail.value === transformationName && !functionNameDetail.isMoreThanTwo
-      )?.length
-    ) {
-      return selectedColumns?.length === 2 ? false : true;
-    } else if (
-      multipleColumnSelected.filter(
-        (functionNameDetail: IMultipleSelectedFunctionDetail) =>
-          functionNameDetail.value === transformationName && functionNameDetail.isMoreThanTwo
-      )?.length
-    ) {
-      return selectedColumns?.length >= 1 ? false : true;
-    } else {
-      return selectedColumns?.length >= 1 ? false : true;
-    }
-  };
-
-  return (
-    <Fragment>
-      <DrawerWidget
-        headingText={T.translate(`${ADD_TRANSFORMATION_PREFIX}.selectColumnPara`)}
-        openDrawer={columnsPopup}
-        showBackIcon={true}
-        closeClickHandler={closeSelectColumnsPopupWithoutColumn}
-        dataTestId="select-column-drawer"
-      >
-        <AddTransformationWrapper>
-          <AddTransformationBodyWrapper>
-            <SelectColumnsList
-              columnsList={columnsList}
-              selectedColumnsCount={selectedColumns.length}
-              setSelectedColumns={setSelectedColumns}
-              dataQuality={dataQualityValue}
-              transformationDataType={transformationDataType}
-              transformationName={transformationName}
-              selectedColumns={selectedColumns}
-            />
-          </AddTransformationBodyWrapper>
-          <AddTransformationButton
-            disabled={enableDoneButton()}
-            color="primary"
-            data-testid="button_done"
-            onClick={closeSelectColumnsPopup}
-          >
-            {T.translate(`${ADD_TRANSFORMATION_PREFIX}.done`)}
-          </AddTransformationButton>
-        </AddTransformationWrapper>
-      </DrawerWidget>
-    </Fragment>
-  );
-}
