+++ app/cdap/components/WranglerGrid/SelectColumnPanel/ColumnTable/RadioInput/index.tsx
@@ -14,5 +14,26 @@
  * the License.
  */
 
-export const MISSING_NULL = 'Missing/Null';
-export const transformationOptions = ['undo', 'redo'];
+import React from 'react';
+import { Radio } from '@material-ui/core';
+import { IRadioInputProps } from 'components/WranglerGrid/SelectColumnPanel/ColumnTable/types';
+
+export default function({
+  selectedColumns,
+  onSingleSelection,
+  columnDetail,
+  columnIndex,
+}: IRadioInputProps) {
+  return (
+    <Radio
+      color="primary"
+      onClick={() => onSingleSelection(columnDetail)}
+      checked={
+        selectedColumns?.filter((column) => column.label === columnDetail.label).length
+          ? true
+          : false
+      }
+      data-testid={`radio-input-${columnIndex}`}
+    />
+  );
+}
